int(math.Round(float64(len(targets)) / 10))

+++ cmd/dnscheck/main.go
@@ -59,7 +59,7 @@ func main() {
 			if len(target) > 1 {
 				wg.Add(1)
 				go dnscheck.Start(target, options.Verbose, wg)
-				if routinescounter == math.Round(float64(len(targets)) / 10)  {
+				if routinescounter == int(math.Round(float64(len(targets)) / 10))  {
 					time.Sleep(5 * time.Second)
 					routinescounter = 0
 				} else {
